basically started over on dungeon gen
i have no idea what i was trying to do with the previous method, it was weird and clunky and would have taken a lot of effort to get a good result from it.

trying a new method where i spawn obstacles around the dungeon, then i spawn the start and final room and try to find a path from the starting room.

didn't have enough time to do the pathfinding itself tho.....